Refactor bolt socket
Move handshake and connect method that can be automatically un-async-ed for
less code duplication.
